airflow_dep/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
lxml
pandas
snowflake-connector-python
apache-airflow-providers-snowflake
snowflake-sqlalchemy
apache-airflow-providers-amazon
requests
urllib3
openpyxl